requirements.txt
confluent-kafka
requests
smolagents
python-dotenv
elasticsearch>=8.0.0,<9.0.0

